+++ AddOns/BetterBlizzFrames/midnight/modules/castbar.lua
@@ -228,38 +228,30 @@ function BBF.SetupBorderOnFrame(frame)
 end
 
 local function UpdateCastTimer(self)
-    if not BBFMIDNIGHT then return end
-    local remainingTime
-    if self.casting or self.reverseChanneling then
-        -- For a cast, we calculate how much time is left until the cast completes
-        remainingTime = self.maxValue - self.value
-    elseif self.channeling then
-        -- For a channel, the remaining time is directly related to the current value
-        remainingTime = self.value
+    if not self.unit then return end
+
+    local duration = UnitCastingDuration(self.unit)
+    if not duration then
+        duration = UnitChannelDuration(self.unit)
     end
 
-    -- If the remaining time is zero or somehow negative, clear the timer
-    if remainingTime then
-        if remainingTime <= 0 then
-            self.Timer:SetText("")
-            return
-        end
-        self.Timer:SetFormattedText("%.1f", remainingTime)
+    if duration then
+        self.Timer:SetFormattedText("%.1f", duration:GetRemainingDuration())
     else
         self.Timer:SetText("")
     end
 end
 
 local function UpdateSparkPosition(castBar)
-    if not BBFMIDNIGHT then return end
-    local val = castBar:GetValue()
-    local minVal, maxVal = castBar:GetMinMaxValues()
-    --local progressPercent = castBar.value / castBar.maxValue
-    if maxVal == 0 then return end
-    local progressPercent = val / maxVal
-    local newX = castBar:GetWidth() * progressPercent
-    castBar.Spark:ClearAllPoints()
-    castBar.Spark:SetPoint("CENTER", castBar, "LEFT", newX, 0)
+    -- if not BBFMIDNIGHT then return end
+    -- local val = castBar:GetValue()
+    -- local minVal, maxVal = castBar:GetMinMaxValues()
+    -- --local progressPercent = castBar.value / castBar.maxValue
+    -- if maxVal == 0 then return end
+    -- local progressPercent = val / maxVal
+    -- local newX = castBar:GetWidth() * progressPercent
+    -- castBar.Spark:ClearAllPoints()
+    -- castBar.Spark:SetPoint("CENTER", castBar, "LEFT", newX, 0)
 end
 
 local function HideChargeTiers(castBar)
@@ -428,7 +420,7 @@ function BBF.ClassicCastbar(castBar, unitType)
     AdjustBorderShieldSize(castBar)
 
     if not castBar.isClassicStyle then
-        castBar:HookScript("OnEvent", function(self)
+        castBar:HookScript("OnEvent", function(self, event)
             self:SetStatusBarTexture(classicCastbarTexture)
             castBar.TextBorder:SetAlpha(0)
             if castBar == PlayerCastingBarFrame then
@@ -503,17 +495,25 @@ function BBF.ClassicCastbar(castBar, unitType)
             if recolorCastbars then
                 local castTexture = self:GetStatusBarTexture()
                 if self.casting then
-                    castTexture:SetVertexColorFromBoolean(
-                        notInterruptible,
-                        castbarColors.colorUninterruptable,
-                        castbarColors.colorStandard
-                    )
+                    if notInterruptible ~= nil then
+                        castTexture:SetVertexColorFromBoolean(
+                            notInterruptible,
+                            castbarColors.colorUninterruptable,
+                            castbarColors.colorStandard
+                        )
+                    else
+                        self:SetStatusBarColor(unpack(castbarColors.standard))
+                    end
                 elseif self.channeling then
-                    castTexture:SetVertexColorFromBoolean(
-                        notInterruptible,
-                        castbarColors.colorUninterruptable,
-                        castbarColors.colorChannel
-                    )
+                    if notInterruptible ~= nil then
+                        castTexture:SetVertexColorFromBoolean(
+                            notInterruptible,
+                            castbarColors.colorUninterruptable,
+                            castbarColors.colorChannel
+                        )
+                    else
+                        self:SetStatusBarColor(unpack(castbarColors.channel))
+                    end
                 else
                     self:SetStatusBarColor(unpack(castbarColors.standard))
                 end
@@ -521,17 +521,25 @@ function BBF.ClassicCastbar(castBar, unitType)
             else
                 local castTexture = self:GetStatusBarTexture()
                 if self.casting then
-                    castTexture:SetVertexColorFromBoolean(
-                        notInterruptible,
-                        castbarColors.defaultUninterruptable,
-                        castbarColors.defaultStandard
-                    )
+                    if notInterruptible ~= nil then
+                        castTexture:SetVertexColorFromBoolean(
+                            notInterruptible,
+                            castbarColors.defaultUninterruptable,
+                            castbarColors.defaultStandard
+                        )
+                    else
+                        self:SetStatusBarColor(unpack(castbarColors.defaultStandard))
+                    end
                 elseif self.channeling then
-                    castTexture:SetVertexColorFromBoolean(
-                        notInterruptible,
-                        castbarColors.defaultUninterruptable,
-                        castbarColors.defaultChannel
-                    )
+                    if notInterruptible ~= nil then
+                        castTexture:SetVertexColorFromBoolean(
+                            notInterruptible,
+                            castbarColors.defaultUninterruptable,
+                            castbarColors.defaultChannel
+                        )
+                    else
+                        self:SetStatusBarColor(unpack(castbarColors.defaultChannel))
+                    end
                 else
                     self:SetStatusBarColor(1.0, 0.7, 0.0, 1)
                 end
@@ -1591,7 +1599,7 @@ local function CastbarOnEvent(self, event)
                 castBarTexture:SetDesaturated(true)
             end
             if castBarRecolorInterrupt and not BBF.playerKickReady then
-                if colors.colorInterruptNotReady then
+                if colors.colorInterruptNotReady and (self.casting or self.channeling) and notInterruptible ~= nil then
                     castBarTexture:SetVertexColorFromBoolean(
                         notInterruptible,
                         colors.colorUninterruptable,
@@ -1602,21 +1610,29 @@ local function CastbarOnEvent(self, event)
                 end
             elseif self.casting then
                 if colors.colorStandard then
-                    castBarTexture:SetVertexColorFromBoolean(
-                        notInterruptible,
-                        colors.colorUninterruptable,
-                        colors.colorStandard
-                    )
+                    if notInterruptible ~= nil then
+                        castBarTexture:SetVertexColorFromBoolean(
+                            notInterruptible,
+                            colors.colorUninterruptable,
+                            colors.colorStandard
+                        )
+                    else
+                        self:SetStatusBarColor(unpack(colors.standard or {1.0, 0.7, 0.0, 1}))
+                    end
                 else
                     self:SetStatusBarColor(unpack(colors.standard or {1.0, 0.7, 0.0, 1}))
                 end
             elseif self.channeling then
                 if colors.colorChannel then
-                    castBarTexture:SetVertexColorFromBoolean(
-                        notInterruptible,
-                        colors.colorUninterruptable,
-                        colors.colorChannel
-                    )
+                    if notInterruptible ~= nil then
+                        castBarTexture:SetVertexColorFromBoolean(
+                            notInterruptible,
+                            colors.colorUninterruptable,
+                            colors.colorChannel
+                        )
+                    else
+                        self:SetStatusBarColor(unpack(colors.channel or {0.0, 1.0, 0.0, 1}))
+                    end
                 else
                     self:SetStatusBarColor(unpack(colors.channel or {0.0, 1.0, 0.0, 1}))
                 end
@@ -1628,14 +1644,14 @@ local function CastbarOnEvent(self, event)
             if castBarTexture then
                 castBarTexture:SetDesaturated(true)
             end
-            if self.casting or self.channeling then
+            if (self.casting or self.channeling) and notInterruptible ~= nil then
                 castBarTexture:SetVertexColorFromBoolean(
                     notInterruptible,
                     colors.defaultUninterruptable,
                     colors.colorInterruptNotReady
                 )
             else
-                self:SetStatusBarColor(unpack(colors.colorInterruptNotReady or {0.7, 0.7, 0.7, 1}))
+                self:SetStatusBarColor(unpack(colors.interruptNotReady or {0.7, 0.7, 0.7, 1}))
             end
             self.changedBarColor = true
         elseif self.changedBarColor then
@@ -1649,36 +1665,48 @@ local function CastbarOnEvent(self, event)
         self:SetStatusBarTexture(classicCastbarTexture)
         if colors.enabled then
             if castBarRecolorInterrupt and not BBF.playerKickReady then
-                castBarTexture:SetVertexColorFromBoolean(
-                    notInterruptible,
-                    colors.colorUninterruptable,
-                    colors.colorInterruptNotReady
-                )
+                if notInterruptible ~= nil then
+                    castBarTexture:SetVertexColorFromBoolean(
+                        notInterruptible,
+                        colors.colorUninterruptable,
+                        colors.colorInterruptNotReady
+                    )
+                else
+                    self:SetStatusBarColor(unpack(colors.interruptNotReady or {0.7, 0.7, 0.7, 1}))
+                end
             elseif self.casting then
-                castBarTexture:SetVertexColorFromBoolean(
-                    notInterruptible,
-                    colors.colorUninterruptable,
-                    colors.colorStandard
-                )
+                if notInterruptible ~= nil then
+                    castBarTexture:SetVertexColorFromBoolean(
+                        notInterruptible,
+                        colors.colorUninterruptable,
+                        colors.colorStandard
+                    )
+                else
+                    self:SetStatusBarColor(unpack(colors.standard or {1.0, 0.7, 0.0, 1}))
+                end
             elseif self.channeling then
-                castBarTexture:SetVertexColorFromBoolean(
-                    notInterruptible,
-                    colors.colorUninterruptable,
-                    colors.colorChannel
-                )
+                if notInterruptible ~= nil then
+                    castBarTexture:SetVertexColorFromBoolean(
+                        notInterruptible,
+                        colors.colorUninterruptable,
+                        colors.colorChannel
+                    )
+                else
+                    self:SetStatusBarColor(unpack(colors.channel or {0.0, 1.0, 0.0, 1}))
+                end
             else
                 self:SetStatusBarColor(unpack(colors.standard or {1.0, 0.7, 0.0, 1}))
             end
         else
             if castBarRecolorInterrupt and not BBF.playerKickReady then
-                if self.casting or self.channeling then
+                if (self.casting or self.channeling) and notInterruptible ~= nil then
                     castBarTexture:SetVertexColorFromBoolean(
                         notInterruptible,
                         colors.defaultUninterruptable,
                         colors.colorInterruptNotReady
                     )
                 else
-                    self:SetStatusBarColor(unpack(colors.colorInterruptNotReady or {0.7, 0.7, 0.7, 1}))
+                    self:SetStatusBarColor(unpack(colors.interruptNotReady or {0.7, 0.7, 0.7, 1}))
                 end
             elseif self.channeling then
                 self:SetStatusBarColor(0, 1, 0)