+++ AddOns/BetterBlizzFrames/CHANGELOG.md
@@ -1,3 +1,29 @@
+# BetterBlizzFrames 1.9.0
+## Prepatch/Midnight
+### New
+- Add Aswog profile (www.twitch.tv/aswog). Thank you for sharing!
+### Tweak
+- Party Range Frame Alpha adjustment setting is now enabled by default in the addon. You'll have to turn it off if you dont want it.
+### Bugfix
+- Fix cast timers
+- fix hide realm name
+- Fix vehicle leave button being hidden when actionbar 1 was hidden
+- Fix custom code section not being clickable in its entire window without text being there
+- Fix misc things
+
+# BetterBlizzFrames 1.8.9h
+## All
+### Bugfix
+- Fix missing locale causing import/export to break
+
+# BetterBlizzFrames 1.8.9g
+## Midnight/Prepatch
+### New
+- New setting: Hide aura tooltips for Target/Focus/Player
+- New setting: Clickthrough Player Auras (you wont be able to click off auras anymore)
+### Bugfix
+- Misc small bugfixes, secret this, bad argument that.
+
 # BetterBlizzFrames 1.8.9f
 ## Midnight/Prepatch
 ### Bugfix